refactor(game-engine): extract turn + foul handlers (S27.8.3)
Slice 3 de S27.8 (refactor monolith `actions.ts`). Trois handlers
clean (aucune dependance interne au dispatcher) extraits dans un
module dedie.

Extraction (`actions/turn-foul-actions.ts`, 172l)
- `handleEndPlayerTurn` : un coach abandonne explicitement la suite
  d'activation d'un joueur (PM=0, gfiUsed=2).
- `handleEndTurn` : changement de tour ; gere les 4 cas (match
  termine, post-TD, blitz kickoff, tour normal) avec reinitialisation
  des flags one-shot (rerollUsedThisTurn, hypnotizedPlayers,
  usedRunningPassThisTurn, usedOnTheBallThisTurn,
  usedMultipleBlockThisTurn).
- `handleFoul` : action FOUL avec gate "1 par tour", consomme
  `mechanics/foul`.

Imports orphelins retires de `actions.ts`
- `canFoul` / `executeFoul` (mechanics/foul)
- `advanceHalfIfNeeded` / `handlePostTouchdown` (core/game-state)

Resultat
- `actions.ts` 2393 -> 2288 lignes (-105 nettes ; cumul -418 vs
  origine 2706 = ~15.4% deja gagne).
- `turn-foul-actions.ts` ajoute a la liste d'exclusion coverage
  (meme quirk source-map v8 que `actions.ts`).
- 4888 tests game-engine passent. Coverage 25.33% maintenue
  (au-dessus du seuil 25%).
- Server : 5 tests pre-existants en echec (engine version mismatch
  0.13.0 -> 0.15.0 dans `sim-engine`, sans rapport avec ce slice).

Reste pour atteindre la cible S27.8 (<= 600 lignes pour `actions.ts`) :
- handleBlock / choice handlers (BlockChoose / PushChoose /
  FollowUpChoose / RerollChoose / DumpOffChoose) — bloques par
  cycle vers `handleBlock` non encore extrait
- handleBlitz
- handleMove / handleNormalMove / handleDodge / handleDodgeRoll /
  handleBallPickup / handleLeap
